requirements.txt
cloudscraper>=1.2.71
beautifulsoup4>=4.12.2
lxml>=4.9.3
fake-useragent>=1.4.0
requests>=2.31.0
python-dotenv>=1.0.0
